% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the Top1000 Shiny App}
\usage{
launch_app()
}
\value{
Opens a Shiny app in your default browser.
}
\description{
This function launches the interactive Shiny application
that allows users to explore how common their name is
compared to others in the Top 1000 dataset.
}
\examples{
if (interactive()) {
  launch_app()
}
}
